+++ sim/core/target.go
@@ -336,7 +336,6 @@ func NewAttackTable(attacker *Unit, defender *Unit, weapon *Item) *AttackTable {
 		// i.e. talents, gear and buffs, NOT base crit and crit from agility!
 		// See https://github.com/magey/classic-warrior/wiki/Attack-table#critical-strike
 		// That means if a character with <2% crit from auras attacks a +3 level target the sim will be wrong.
-		// The chance of that being the case once bosses are +3 in SoD should be very small though.
 		// Most (all?) affected specs have crit in their talents to begin with.
 		if (defender.Level - attacker.Level) >= 3 {
 			table.MeleeCritSuppression += 0.018